Updating order of the mpi4py lecture so communicator stuff comes at the end not the beginning
