+++ lectures/dev-mpi4py/example2.py
@@ -1,23 +1,22 @@
-"""Creating a Cartesian Communicator"""
+"""A simple blocking Send/Recv pair"""
 
 import sys
+import numpy
 from mpi4py import MPI
 
-def main(parent):
-    """Create a 2-d Cartesian communicator from parent communicator"""
-
-    dims = MPI.Compute_dims(parent.size, 2)
-
-    comm = parent.Create_cart(dims, periods = [True, False])
+def main(comm, sz):
+    """Send a message between ranks 0 and 1"""
     rank = comm.Get_rank()
-    coords = comm.Get_coords(rank)
-    upx = comm.Shift(0, 1)
-    upy = comm.Shift(1, 1)
 
-    out = "Rank{:2d} coords{:2d} {:2d} upx(src,dst) {} upy(src,dst) {}\n"\
-                         .format(rank, coords[0], coords[1], upx, upy)
-    sys.stdout.write(out)
+    if rank == 0:
+        msg = numpy.ones(sz, numpy.double)
+        comm.Ssend([msg, sz, MPI.DOUBLE], dest = 1, tag = 999)
+    elif rank == 1:
+        msg = numpy.zeros(sz, numpy.double)
+        comm.Recv([msg, sz, MPI.DOUBLE], source = 0, tag = 999)
+        sys.stdout.write("Rank 1 received {}".format(msg))
 
 if __name__ == "__main__":
-    """Execute in MPI.COMM_WORLD"""
-    main(MPI.COMM_WORLD)
+
+    sz = 4
+    main(MPI.COMM_WORLD, sz)
