--- lectures/dev-mpi4py/images
@@ -1,1 +1,0 @@
-../images
